git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@19865 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/gui/swing/jdgui/JDGui.java
@@ -254,6 +254,7 @@ public class JDGui extends SwingGui {
                 Application.getResource("/tmp/update/self/JDU").mkdirs();
                 new Thread() {
                     public void run() {
+                        if (UpdateController.getInstance().getHandler() == null) return;
                         if (Application.getResource("JDownloader.jar").lastModified() < new Date(2013 - 1900, 1, 10, 10, 0).getTime()) {
                             try {
                                 Files.deleteRecursiv(Application.getResource("cfg/versioninfo/JDU"));
